pyproject.toml
[project]
name = "adapt3r"
version = "2.0.0"
description = "Official implementation of Adapt3R"
authors = [
    { name = "Albert Wilcox", email = "albertwilcox@gatech.edu" },
    { name = "Mohamed Ghanem", email = "mghanem8@gatech.edu" },
    { name = "Atharva Mete", email = "amete7@gatech.edu" }
]
dependencies = [
    "easydict",
    "einops==0.8.0",
    "h5py==3.11.0",
    "hydra-core==1.3.2",
    "imageio==2.34.1",
    "matplotlib==3.9.0",
    "numpy==1.26.4",
    "positional-encodings==6.0.3",
    "torch==2.4.0",
    "tqdm==4.66.4",
    "vector-quantize-pytorch==1.14.24",
    "wandb==0.17.0",
    "cmake==3.29.3",
    "diffusers==0.30.0",
    "pyinstrument==4.6.2",
    "moviepy==2.1.1",
    "tokenizers==0.19.1",
    "opencv-python==4.10.0.84",
    "classifier-free-guidance-pytorch==0.6.10",
    "torchtyping==0.1.4",
    "ema-pytorch==0.5.2",
    "thop==0.1.1-2209072238",
    "bddl==1.0.1",
    "future==0.18.2",
    "gym==0.25.2",
    "cloudpickle==2.1.0",
    "robosuite==1.4.1",
    "transformers",
    "huggingface_hub==0.24.6",
    "natsort==8.4.0",
    "open3d==0.19.0",
    "gymnasium==1.0.0",
    "submitit==1.5.3",
    "clip",
    "robomimic",
]
requires-python = "==3.10.*"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
]

[project.urls]
Repository = "https://github.com/pairlab/Adapt3R/"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.isort]
known-first-party = ["adapt3r"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["torch.*", "numpy.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=adapt3r"

[tool.hatch.build.targets.wheel]
packages = ["adapt3r"] 

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git", rev = "main" }
robomimic = { git = "https://github.com/ARISE-Initiative/robomimic.git", rev = "59ecd4048b91595b61cbcc2f3e4d9d666e4d4c04" }
